+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Focus and keyDown events propagation stopped for component being used inside `terra-compact-interactive-list` cells.
+
 ## 6.88.0 - (March 8, 2024)
 
 * Changed